# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "rentNest Gateway API",
    "version": "1.0.0",
    "description": "OpenAPI specification for the rentNest API gateway — aligned with provided TypeScript models, DTOs and EntitySchemas."
  },
  "servers": [
    { "url": "http://localhost:2000", "description": "Local gateway" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "Authorization: Bearer <accessToken>" },
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "refreshToken" }
    },
    "schemas": {
      "UserRole": { "type": "string", "enum": ["user", "moderator", "admin"] },
      "UserStatus": { "type": "string", "enum": ["active", "banned"] },
      "IUser": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "email": { "type": "string", "format": "email" },
          "passwordHash": { "type": "string", "nullable": true },
          "googleID": { "type": "string", "nullable": true },
          "name": { "type": "string", "nullable": true },
          "photo": { "type": "string", "nullable": true },
          "role": { "$ref": "#/components/schemas/UserRole" },
          "status": { "$ref": "#/components/schemas/UserStatus" },
          "ip_address": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time", "readOnly": true },
          "updatedAt": { "type": "string", "format": "date-time", "nullable": true, "readOnly": true }
        },
        "required": ["id", "email", "role", "status", "createdAt"]
      },
...
      "ReservationStatus": {
        "type": "string",
        "enum": ["pending", "confirmed", "declined", "cancelled", "paid", "checked_in", "checked_out"]
      },
      "FavoriteStatus": { "type": "string", "enum": ["uploaded", "deleted"] },
      "ReviewStatus": { "type": "string", "enum": ["uploaded", "deleted"] },
      "PropertyStatus": { "type": "string", "enum": ["active", "deleted"] },
...
      "Favorite": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "readOnly": true },
          "userId": { "type": "integer" },
          "propertyId": { "type": "integer" },
          "status": { "$ref": "#/components/schemas/FavoriteStatus" },
          "createdAt": { "type": "string", "format": "date-time", "readOnly": true }
        }
      },
...
      "Property": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "ownerId": { "type": "integer", "readOnly": true },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "city": { "type": "string" },
          "region": { "type": "string", "nullable": true },
          "street": { "type": "string", "nullable": true },
          "pricePerNight": { "type": "number" },
          "currency": { "type": "string", "nullable": true },
          "bedrooms": { "type": "integer" },
          "bathrooms": { "type": "integer" },
          "area": { "type": "number", "nullable": true },
          "amenities": { "type": "array", "items": { "type": "string" }, "nullable": true },
          "photos": { "type": "array", "items": { "type": "string", "format": "uri" }, "nullable": true },
          "availableFrom": { "type": "string", "format": "date-time", "nullable": true },
          "availableTo": { "type": "string", "format": "date-time", "nullable": true },
          "status": { "$ref": "#/components/schemas/PropertyStatus" },
          "createdAt": { "type": "string", "format": "date-time", "readOnly": true },
          "updatedAt": { "type": "string", "format": "date-time", "nullable": true, "readOnly": true }
        }
      },
...
      "Reservation": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "propertyID": { "type": "integer" },
          "tenantID": { "type": "integer" },
          "ownerId": { "type": "integer", "readOnly": true },
          "guests": { "type": "integer" },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "status": { "$ref": "#/components/schemas/ReservationStatus" },
          "createdAt": { "type": "string", "format": "date-time", "readOnly": true },
          "updatedAt": { "type": "string", "format": "date-time", "nullable": true, "readOnly": true }
        }
      },
...
      "Review": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "readOnly": true },
          "rating": { "type": "integer" },
          "comment": { "type": "string" },
          "bookingID": { "type": "integer" },
          "propertyID": { "type": "integer" },
          "senderID": { "type": "integer" },
          "propertyOwnerId": { "type": "integer" },
          "status": { "$ref": "#/components/schemas/ReviewStatus" },
          "createdAt": { "type": "string", "format": "date-time", "readOnly": true },
          "updatedAt": { "type": "string", "format": "date-time", "nullable": true, "readOnly": true }
        }
      },
...
      "TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "expiresIn": { "type": "integer", "nullable": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/auth/google/callback": {
      "get": {
        "tags": ["Auth"],
        "summary": "Google OAuth2 callback",
        "responses": { "302": { "description": "Redirect to client with token" }, "401": { "description": "Auth failed" } }
      }
    },
    "/api/auth/refresh": {
      "get": {
        "tags": ["Auth"],
        "summary": "Refresh access token (reads refreshToken cookie)",
        "responses": {
          "200": {
            "description": "New access token",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenResponse" } } }
          },
          "401": { "description": "Invalid or missing token" }
        }
      }
    },
...
    "/api/user": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Array of users", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/IUser" } } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/favorites": {
      "get": {
        "tags": ["Favorites"],
        "summary": "Get favorites for current user",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Array of favorites", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Favorite" } } } } }, "401": { "description": "Unauthorized" } }
      }
    },
...
    "/api/properties": {
      "get": {
        "tags": ["Properties"],
        "summary": "List properties (supports query params for pagination & filters)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "city", "in": "query", "schema": { "type": "string" } },
          { "name": "minPrice", "in": "query", "schema": { "type": "number" } },
          { "name": "maxPrice", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": { "200": { "description": "Array of properties", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Property" } } } } } }
      }
    },
...
    "/api/reservations": {
      "get": {
        "tags": ["Reservations"],
        "summary": "Get all reservations (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Array of reservations", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Reservation" } } } } } }
      }
    },
...
    "/api/reviews": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Get all reviews (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Array of reviews", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Review" } } } } } }
      }
    },
...
    "/api/search/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Search properties",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Search results", "content": { "application/json": { "schema": { "type": "object", "properties": { "total": { "type": "integer" }, "items": { "type": "array", "items": { "$ref": "#/components/schemas/Property" } } } } } } } }
      }
    }
  }
}